+++ packages/tswebextension/src/lib/common/message-constants.ts
@@ -23,4 +23,5 @@ export enum MessageType {
     CloseAssistant = 'closeAssistant',
     AssistantCreateRule = 'assistantCreateRule',
     SaveCssHitsStats = 'saveCssHitsStats',
+    GetRemoveParamUrl = 'getRemoveParamUrl',
 }