+++ packages/tswebextension/src/lib/mv3/content-script/index.ts
@@ -1,3 +1,5 @@
+import { initRemoveParamBridge } from '../../common/content-script/remove-param-handler';
+
 import { CookieController } from './cookie-controller';
 import { CosmeticController } from './cosmetic-controller';
 
@@ -6,3 +8,5 @@ cosmeticController.init();
 
 const cookieController = new CookieController();
 cookieController.init();
+
+initRemoveParamBridge();