+++ packages/tswebextension/test/lib/mv3/background/services/document-blocking-service.test.ts
@@ -69,6 +69,7 @@ describe('DocumentBlockingService', () => {
             stealthModeEnabled: false,
             gpcScriptUrl: '/gpc.js',
             hideDocumentReferrerScriptUrl: '/hide-referrer.js',
+            removeParamScriptUrl: '/remove-param.js',
             stealth: {
                 blockChromeClientData: false,
                 hideReferrer: false,